Add an example Global app that searchs appliances by tag via GraphQL
The global app receives the Tenant GraphQL API url and the necessary
auth token via props when it is instantiated, and proceeds to query for
the appliance list and the existing appliance tags.
Selecting one of the existing tags triggers a new query that searches
only appliances that have the chosen tag.

Signed-off-by: Davide Briani <davide.briani@secomind.com>